+++ src/App.tsx
@@ -1,8 +1,8 @@
-import axios from 'axios';
 import * as React from 'react';
 import './App.css';
 import Main from './container/main/main';
-import { filterByClothingType } from './utilities/helper';
+// import MainProvider from './container/mainProvider/mainProvider';
+import { MainContext } from './container/mainProvider/mainProvider';
 
 export interface IAddOutfitPayload {
   headwearSelected  : IClothing;
@@ -28,15 +28,6 @@ export interface IOutfit {
   top       : IClothing[],
   bottom    : IClothing,
   shoes     : IClothing
-}
-
-interface IState {
-  clothings : IClothing[],
-  outfits   : IOutfit[],
-  headwear  : IClothing[];
-  top       : IClothing[];
-  bottom    : IClothing[];
-  shoes     : IClothing[];
 }
 
 export const CLOTHING_TYPES = [
@@ -71,50 +62,7 @@ export const initOutfit = {
   shoes     : initClothing
 }
 
-export default class App extends React.Component < {}, IState > {
-  constructor(props: {}){
-    super(props)
-    this.state = {
-      clothings : [initClothing],
-      outfits   : [initOutfit],
-      headwear  : [initClothing],
-      top       : [initClothing],
-      bottom    : [initClothing],
-      shoes     : [initClothing]
-    }
-  }
-  public componentDidMount = () => {
-    // fetch all data and store
-    axios.all([
-      axios.get('http://localhost:4000/clothings/type/shoes'),
-      axios.get('http://localhost:4000/clothings/type/shirts'),
-      axios.get('http://localhost:4000/clothings/type/pants'),
-      axios.get('http://localhost:4000/clothings/type/polos'),
-      axios.get('http://localhost:4000/clothings/type/jackets'),
-      axios.get('http://localhost:4000/clothings/type/sweaters'),
-      axios.get('http://localhost:4000/clothings/type/hoodies'),
-      axios.get('http://localhost:4000/outfits/')
-    ])
-    .then(axios.spread((shoes, shirts, pants, polos, jackets, sweaters, hoodies, outfits) => {
-      const clothings = [
-        ...shoes.data.data,
-        ...shirts.data.data,
-        ...pants.data.data,
-        ...polos.data.data,
-        ...jackets.data.data,
-        ...sweaters.data.data,
-        ...hoodies.data.data
-      ]
-      this.setState({
-        clothings,
-        headwear      : filterByClothingType( clothings , ['hat']),
-        top           : filterByClothingType( clothings , ['jackets', 'sweaters', 'shirts', 'polos']),
-        bottom        : filterByClothingType( clothings , ['pants']),
-        shoes         : filterByClothingType( clothings , ['shoes']),
-        outfits : outfits.data.data
-      })
-    }));
-  }
+export default class App extends React.Component < {}, {} > {
   public render() {
     const {
       clothings,
@@ -123,20 +71,24 @@ export default class App extends React.Component < {}, IState > {
       top,
       bottom,
       shoes
-    } = this.state
-
-    const mainProps = {
-      clothings,
-      outfits,
-      headwear,
-      top,
-      bottom,
-      shoes
-    }
+    } = this.context
     return (
       <div className="container-fluid">
-        <Main {...mainProps} />
+        <Main
+          {
+            ...{
+              clothings,
+              outfits,
+              headwear,
+              top,
+              bottom,
+              shoes
+            }
+          }
+        />
       </div>
     );
   }
 }
+
+App.contextType = MainContext;
